struct CColor becomes typedef juce::Colour
escape CColor class gets retired; and is now just a straight
typedef to juce::Colour

+++ src/common/SkinColors.h
@@ -24,7 +24,7 @@ struct Color
     static std::vector<Color> getAllColors();
 
     std::string name;
-    int r, g, b, a;
+    uint8_t r, g, b, a;
 };
 } // namespace Skin
 } // namespace Surge
